MNMS-462 Refactor: BookingCommandService 리팩토링
BookingStatusService 생성
